chore(tests): rename support module gen to graphgen
Renamed the test support module from `gen` to `graphgen` to better reflect its purpose of generating graph structures for fuzz testing, and updated the corresponding import in the fuzz test file.

Auto-committed-on: dragonfly
