+++ tests/support/mod.rs
@@ -6,4 +6,4 @@
 
 #![allow(dead_code)]
 
-pub mod gen;
+pub mod graphgen;
